requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.1
jinja2>=3.1.0
pytest>=7.4.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
requests>=2.31.0
pillow>=10.0.0
rich>=13.0.0
typing-extensions>=4.8.0
flask>=3.0.0